Complete
10 level is done

+++ JavaRushTasks/1.JavaSyntax/src/com/javarush/task/task10/task1006/Solution.java
@@ -9,9 +9,9 @@ public class Solution {
         short b = (short) 45;
         char c = (short) 'c';
         short s = (short) 1005.22;
-        int i = (short) 150000;
-        float f = (short) 4.10f;
-        double d = (short) 1.256d;
+        int i = 150000;
+        float f =  4.10f;
+        double d =  1.256d;
         double result = (f * b) + (i / c) - (d * s) + 562.78d;
         System.out.println("result: " + result);
     }
